Fix: Expanding and Correcting Central Differences Function
Alterations:
-Changing the order the arguments of the negative binomial log
likelihood function. This is useful for defining partial functions
based on this function.
-Correcting and refactoring the approx_Taylor_expansion function ,
so that multidimensional arrays are allowed to be used and there's
a clear separation of functions using first and second order central
differences.
-The original central differences function was converted into the
fst_order_central_differences function. To clearly indicate that
this corresponded to first-order central differences.
-A new function was created that approximates the second order
derivative (snd_order_central_differences function).