Fix: Correcting the Log-Likelihood of the Negative Binomial Distribution
Alters the function log_likelihood_neg_binom so that if the combination of
the means and the observations renders a 3D array, it returns an 1D array
where the values correspond to the sums of the last two dimensions
(observations and features) in the 3D array. If the combination of the means
and the observations renders a 2D array, then it returns a 1D array, where the
values correspond to the sums of the last dimension (features).