Fix: Newton-Raphson method using sparse diagonal matrix
Corrects:
-log_likelihood_neg_binom function: it calculates the negative
binomial of the entire input, instead of calculating the likeli-
hood until it forms an array.

-newton_raphson_method function: it specifies the exact format
of the matrices being used, and adds the number of offsets used
on each side in the solve_banded function. With these
alterations, it can actually solve both positive definite and
semi-positive definite matrix equations.