Feat: Restructuring and Adding Functions to Allow a Parameterized GMRF
Additions:
-build_gmrf_precision_mat was created in order to facilitate the
creation of the precision matrix of a model that corresponds to
the sum of an intercept, a fixed-effects component and a first
-order Random-Walk, with parameters that can be chosen when
using this function.

-gmrf_density was created in order to have a function that
calculates the density of the gmrf (which corresponds to a
Multivariate Normal Distribution).

-create_gmrf_density_func was created in order to not have to
not have to recompute the determinant of the same precision
matrix multiple times. This generates a density function for a
given precision matrix.

Alterations:
-log_likelihood_neg_binom was altered in order to accomodate the
fact that although the gmrf corresponds to a vector of 2n+1
elements only the last n are used to generated the observations.

-newton_raphson_method was altered in order to incorporate the
Cholesky Decomposition has we can make the assumption of
symmetric positive-definite matrices. Some of the arguments were
also adjusted to the new definition of GMRF that has mean 0.

-approx_marg_post_of_theta was altered in order to accomodate the
changes made to the rest of the code.