Tests: New Sampler and Adapting Tests
Alterations to the sampler and the tests were made to accomodate
the new model.

Sampler:
-creating_IGMRF_Q was created to easily construct the precision
matrix of a first-order Random Walk.

-sampling_IGMRF was created to sample from a IGMRF given a
precision matrix. Since this one is rank deficient, the sampling
has to be done through the eigenvectors associated with non-zero
eigenvalues.

-sampling_global_GMRF was created to sample from a model that
incorporates an intercept, a first-order Random-Walk and a fixed
effects component.

-sampling_synthetic_data generates the observations from a
sample of a GMRF.

Adaptations:
The entire class of TestConditionalDensities was adapted to
incorporate the changes in the model structure and the sampler.