Refac: Newton-Raphson and the l-bfgs methods
 -Readjusts the Newton-Raphson method to obtain the mode of the
distribution by using a "Newton-CG" minimizer on the GPU and to
calculate the precision matrix of the Gaussian approximation
using the diagonal of the Hessian matrix at that point,
calculated using torch autograd.
 -Eliminates my implementation of finite differences functions
and taylor expansion functions.
 -Creates a constraint enforcer function that increases the
accuracy of the mode obtained through the Newton-Raphson methdo
(Fisher Scoring method).
 -Converting the previous l_bfgs function into the
calc_mode_of_marg_post_theta as it is a more descriptive name.
This function though still implements the L-BFGS method, now use
a GPU minimizer making it faster. This minimizer however does not
allow for box-constraints. As way to correct this, I include the
a bouding_func as an argument that allows a function to map the
provided thetas by the minimizers into the bounded space that the
thetas actually belong to.