Fix: Restructuring, fixing internal minimizer and Dask
utils:
  -adding the generalized determinant and the tensor creator
functions;
  -creating a wrapper for pytorch tensors to be able to be used
as backend for dask chunks;
  -correcting a bug in dask array, that prevent the
concatenation of non-numpy arrays.

inla_TD:
  -adaptation to the creation of the utils file;
  -fixing the minimizer in the p_x_given_y_theta, so that box
constraints that possibly prevent overflow (depending on the
link function) can be applied;
  -changing the error message in the explorer function, so that
the parameters that caused the error message can be seen.

distributions_TD:
  -adaptation to the creation of utils file;
  -adapting the negative binomial likelihood function to use
observations that are stored in a (possibly out-of-memory) dask
array;
  -adapting the function that creates the Random Walk precision
matrix so that it creates the precision matrix directly in
pytorch in a dense format.